Vim-Adaptor: package refactor to put Heat related placeholder objects as a sub-package of wrapper.openstack package. License header polishing
