+++ vim-adaptor/adaptor/src/main/java/sonata/kernel/vimadaptor/wrapper/openstack/heat/HeatNet.java
@@ -24,30 +24,76 @@
  * 
  */
 
-package sonata.kernel.vimadaptor.commons.heat;
+package sonata.kernel.vimadaptor.wrapper.openstack.heat;
 
-import java.util.ArrayList;
-import java.util.Collections;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class HeatModel {
+public class HeatNet {
 
-  private ArrayList<HeatResource> resources;
 
-  public HeatModel() {
-    this.resources = new ArrayList<HeatResource>();
+  @JsonProperty("net_name")
+  private String netName;
+
+  @JsonProperty("net_id")
+  private String netId;
+
+  @JsonProperty("subnet_name")
+  private String subnetName;
+
+  @JsonProperty("subnet_id")
+  private String subnetId;
+
+  @JsonProperty("segmentation_id")
+  private int segmentationId;
+
+  private String cidr;
+
+  public String getNetName() {
+    return netName;
   }
 
-  public void addResource(HeatResource res) {
-    this.resources.add(res);
+  public String getNetId() {
+    return netId;
   }
 
-  public void prepare() {
-    Collections.sort(resources);
-    return;
+  public String getSubnetName() {
+    return subnetName;
   }
 
-  public ArrayList<HeatResource> getResources() {
-    return this.resources;
+  public String getSubnetId() {
+    return subnetId;
+  }
+
+  public int getSegmentationId() {
+    return segmentationId;
+  }
+
+  public String getCidr() {
+    return cidr;
+  }
+
+  public void setNetName(String netName) {
+    this.netName = netName;
+  }
+
+  public void setNetId(String netId) {
+    this.netId = netId;
+  }
+
+  public void setSubnetName(String subnetName) {
+    this.subnetName = subnetName;
+  }
+
+  public void setSubnetId(String subnetId) {
+    this.subnetId = subnetId;
+  }
+
+  public void setSegmentationId(int segmentationId) {
+    this.segmentationId = segmentationId;
+  }
+
+  public void setCidr(String cidr) {
+    this.cidr = cidr;
   }
 
 }
